feat: add vitals management functionality with CRUD operations
- Implemented vitals controller with create, read, update, and delete functionalities.
- Added Vitals model to define the schema for vital records.
- Created routes for vitals management including endpoints for creating, retrieving, updating, and deleting vitals.

feat: add diagnosis management functionality

- Introduced Diagnosis model and controller for managing diagnoses.
- Implemented CRUD operations for diagnoses.
- Created routes for diagnosis management including endpoints for creating, retrieving, updating, and deleting diagnoses.

feat: add findings management functionality

- Introduced Findings model and controller for managing findings.
- Implemented CRUD operations for findings.
- Created routes for findings management including endpoints for creating, retrieving, updating, and deleting findings.

feat: add investigation management functionality

- Introduced Investigation model for managing investigations.
- Implemented schema for investigations.

feat: add symptoms management functionality

- Introduced Symptoms model and controller for managing symptoms.
- Implemented CRUD operations for symptoms.
- Created routes for symptoms management including endpoints for creating, retrieving, updating, and deleting symptoms.

feat: add properties management for diagnosis, findings, and symptoms

- Introduced models for DiagnosisProperties, FindingsProperties, and SymptomsProperties.
- Implemented schema for properties management.
- Created routes for managing properties related to diagnoses, findings, and symptoms.

+++ controllers/clinic.js
@@ -132,8 +132,8 @@ export const getClinicsByDoctorOrStaffId = async (req, res) => {
 
     if (!clinics || clinics.length === 0) {
       return res
-        .status(404)
-        .json({ success: false, message: "No clinics found" });
+        .status(200)
+        .json({ success: false, message: "No clinics found", clinics: [] });
     }
 
     res.status(200).json({ success: true, clinics });
@@ -191,6 +191,24 @@ export const getClinic = async (req, res) => {
       message: "Internal server error",
       error: error.message,
     });
+  }
+};
+
+export const getClinicById = async (req, res) => {
+  const { clinicId } = req.params;
+
+  try {
+    const clinic = await Clinic.findById(clinicId);
+    if (!clinic) {
+      return res
+        .status(404)
+        .json({ success: false, message: "Clinic not found" });
+    }
+
+    res.status(200).json({ success: true, clinic });
+  } catch (error) {
+    console.error("Error fetching clinic:", error);
+    res.status(500).json({ success: false, message: "Internal server error" });
   }
 };
 
@@ -253,6 +271,25 @@ export const updateClinic = async (req, res) => {
 
 export const deleteClinic = async (req, res) => {
   const { id } = req.params;
+
+  const userId = req.user.id;
+  const userRole = req.user.role;
+
+  // check if the doctor has last clinic and try to delete the clinic
+  if (userRole === "staff") {
+    return res.status(403).json({
+      success: false,
+      message: "Staff cannot delete clinics",
+    });
+  }
+
+  const allClinics = await Clinic.find({ doctorId: userId });
+  if (allClinics.length === 1) {
+    return res.status(403).json({
+      success: false,
+      message: "You cannot delete the last clinic",
+    });
+  }
 
   try {
     const clinic = await Clinic.findByIdAndDelete(id);
